Renamed no-attr-in-components to no-attrs-in-components to reflect the actual name of attrs in the rule name.
