+++ lib/rules/index.js
@@ -36,5 +36,5 @@ module.exports = {
   'no-unnecessary-concat': require('./lint-no-unnecessary-concat'),
   'no-partial': require('./lint-no-partial'),
   'invocable-blacklist': require('./lint-invocable-blacklist'),
-  'no-attr-in-components': require('./lint-no-attr-in-components'),
+  'no-attrs-in-components': require('./lint-no-attrs-in-components'),
 };
