+++ lib/rules/lint-no-attrs-in-components.js
